Gapminder Income Inequality
The Gini index measures the extent to which the distribution of income (or in some cases, consumption expenditure) among individuals or households with an economy deviates from a perfectly equal distribution.  A Lorenz curve plots the cumulative percentages of the total income received against the cumulative number of  recipients, starting with the poorest individual or househould.  The Gini index measures the area between the Lorenz curve and a hypothetical line of absolute equality, expressed as a percentage of max area under the line.  Thus a Gini index of 0 represents perfect equality while 100 perfect inequality.