Still typing via my phone
So only a few problems a day...

I wish I had not pawned my laptop for shelter but if I had not, well they probably would have stolen it.  Knocking me down will get you nothing... And it won't build for me because I did not start in LA or NYC or even DC, although DC might be trying to own me, although I signed with Kiran with Blackstraw in 2022.  All have an interest in leaving me standing. None have an interest in giving me any love or comfort or even building for me.