Added a problem from 4-4
Still alive, and laptopless
But was able to wash my laundry and pay for my vpn subscription after a small gift from a stranger. I drank a Shasta root beer... This is the life the neobandists want for women.  Sonlife and sunscreen isn't my sim or my course but other people love to make my life hell because they hate me and don't care that I would prefer to ho back and be an engineering manager for NGA or something.  F all the managers that say pick a monotheistic religion before we let you do anything - I am an atheist.  I have taken shelter among muslims, bit the devout there are just as dangerous.